requirements.txt
pandas
duckdb
pyarrow
requests
pytest
python-dateutil
